requirements.txt
click==8.0.1
coverage==6.0
dataclasses==0.8
Flask==2.0.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pip==21.3.1
pydicom==2.2.2
setuptools==58.0.4
tomli==1.2.2
typing-extensions==3.10.0.2
Werkzeug==2.0.1
wheel==0.37.0
zipp==3.6.0

torch==1.4.0
torchvision==0.2.0
scikit-learn==0.23.1
scipy
numpy
Pillow
tqdm
mock
statsmodels
scikit-image
imageio
gitpython
lifelines
